requirements.txt
Flask==3.0.3
mysql-connector-python==9.1.0
reportlab==4.2.2
Werkzeug==3.0.4
python-dotenv==1.0.1
bcrypt==4.1.3
Flask-WTF==1.2.1
WTForms==3.1.2

